refactor(server): #125 — extract status/health handlers into routes/status.rs
