+++ src/server/routes/identity.rs
@@ -21,8 +21,8 @@ use crate as x0x;
 use super::super::state::AppState;
 use super::super::{
     api_error, bad_request, has_withdrawn_same_stable_group_record, parse_optional_json,
-    ApiResponse,
 };
+use super::status::ApiResponse;
 
 /// POST /agent/sign request body — a caller payload to sign with the
 /// agent's ML-DSA-65 secret key under a mandatory external domain.
